Break views (and its tests) module down into multiple modules
